requirements.txt
asgiref==3.4.1
async-timeout==4.0.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.5
cryptography==40.0.2
Django==3.2.18
django-comment-migrate==0.1.7
django-mirage-field==1.3.0
django-proxy==1.2.2
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
packaging==21.3
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.1.1
pytz==2023.3.post1
redis==4.3.5
requests==2.27.1
sqlparse==0.4.4
tqdm==4.64.1
typing_extensions==4.1.1
urllib3==1.26.17
zipp==3.6.0
